chore: move test route configuration to the correct place
